+++ src/common/POLFields/PhysicalFields/FieldMaterialSupplier.js
@@ -1,22 +1,27 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import { FieldSelection } from '@folio/stripes-acq-components';
+import { FieldOrganization } from '../../POFields';
 
-const FieldMaterialSupplier = ({ vendors, disabled }) => {
+const FieldMaterialSupplier = ({ change, dispatch, disabled, materialSupplierId }) => {
   return (
-    <FieldSelection
-      dataOptions={vendors}
+    <FieldOrganization
+      dispatch={dispatch}
+      change={change}
       labelId="ui-orders.physical.materialSupplier"
       name="physical.materialSupplier"
       disabled={disabled}
+      id={materialSupplierId}
+      required={false}
     />
   );
 };
 
 FieldMaterialSupplier.propTypes = {
-  vendors: PropTypes.arrayOf(PropTypes.object).isRequired,
+  dispatch: PropTypes.func.isRequired,
+  change: PropTypes.func.isRequired,
   disabled: PropTypes.bool,
+  materialSupplierId: PropTypes.string,
 };
 
 FieldMaterialSupplier.defaultProps = {