+++ test/bigtest/interactors/order-edit-page.js
@@ -4,6 +4,7 @@ import {
   isPresent,
   text,
   value,
+  fillable,
 } from '@bigtest/interactor';
 
 import { OptionListInteractor } from '@folio/stripes-acq-components/test/bigtest/interactors';
@@ -19,11 +20,6 @@ import { TIMEOUT } from './const';
 @interactor class OrderTypeSelect {
   static defaultScope = 'select[name="orderType"]';
   value = value();
-}
-
-@interactor class VendorSelect {
-  button = new Button('[name="vendor"]');
-  options = new OptionListInteractor('#sl-po-vendor');
 }
 
 @interactor class OrderTemplate {
@@ -49,7 +45,7 @@ export default interactor(class OrderEditPage {
   renewalsAccordion = isPresent('#renewals');
   orderTypeSelect = new OrderTypeSelect();
   createOrderButton = new Button('#clickable-create-new-purchase-order');
-  vendorSelect = new VendorSelect();
+  fillVendor = fillable('[name="vendor"]');
   addNoteButton = new Button('[data-test-add-note-button]');
   removeNoteButton = new Button('[data-test-remove-note-button]');
   notes = collection('[name*=notes]');