+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## (IN PROGRESS)
 
 ### Stories
+* [UIOR-506](https://issues.folio.org/browse/UIOR-506) Replace org selection components with organization lookup
 * [UIOR-491](https://issues.folio.org/browse/UIOR-491) Cloning a PO and its POL(s)
 * [UIOR-457](https://issues.folio.org/browse/UIOR-457) Ability to "Reopen" and once again "Close" purchase orders
 * [UIOR-509](https://issues.folio.org/browse/UIOR-509) Move record action buttons into "Action" button UX pattern